Commands are now part of a Commander Type which holds Driver. Tag renamed to Device.
This aligns better with the spec's naming schemes and improves method
overview in the documentation. It makes sense that a Commander is the one
in charge to perform operations such Select().

It makes sense that we don't keep a Driver global variable, as this is something
only the Commander needs to know. A Device.Setup(CommandDriver) method
is provided which eases the commander setup. In general, things are more
intuitive.

byte_utils.go methods have been made private. This is not a module to play with
bytes and we offer all the stuff to get going without such functions.