Reduce static Tag MLe MLc values to 128 (this is more sensible)
256 means the maximum supported ReadBinary payload is that. But, per RATs,
the maximum frame size supported by the ISO 14443 standard and readers
(like my phone) is 256. So 128 seems more conservative and gives some space
for CRCs, the rest of the APDU etc. This should be customizable though.

I'm not sure what happens if ReadBinary asks for more bytes than a frame
can support, but probably nothing good.